feat: Added test. Moving function to _x namespace.
