+++ reflex/__init__.py
@@ -120,7 +120,7 @@ _ALL_COMPONENTS = [
 _MAPPING = {
     "reflex.experimental": ["_x"],
     "reflex.admin": ["admin", "AdminDash"],
-    "reflex.app": ["app", "App", "UploadFile", "asset"],
+    "reflex.app": ["app", "App", "UploadFile"],
     "reflex.base": ["base", "Base"],
     "reflex.compiler": ["compiler"],
     "reflex.components": _ALL_COMPONENTS,
