+++ reflex/experimental/__init__.py
@@ -7,6 +7,7 @@ from reflex.components.sonner.toast import toast as toast
 
 from ..utils.console import warn
 from . import hooks as hooks
+from .assets import asset as asset
 from .layout import layout as layout
 from .misc import run_in_thread as run_in_thread
 
@@ -20,4 +21,5 @@ _x = SimpleNamespace(
     progress=progress,
     run_in_thread=run_in_thread,
     toast=toast,
+    asset=asset,
 )
