+++ tests/experimental/custom_script.js
@@ -1,0 +1,1 @@
+const test = "inside custom_script.js";
